backend/requirements.txt
django==5.0.4
djangorestframework==3.14.0
django-cors-headers==4.3.1
python-dotenv==1.0.1
firebase-admin==6.4.0
pydantic==2.6.4
openpyxl==3.1.2
dependency-injector==4.49.0
pytest==8.1.1
pytest-django==4.8.0
pytest-cov==4.1.0
pytest-mock==3.12.0
gunicorn==21.2.0
django-ratelimit==4.1.0
drf-spectacular==0.27.2
whitenoise==6.6.0
python-dotenv==1.0.1
